# fhir/static/swagger.json
{
  "info": {
    "description": "",
    "version": "1.0.0",
    "title": "Fhir"
  },
  "schemes": [
    "http","https"
  ],
  "tags": [

  ],
  "paths": {
    "/api/v1/login": {
      "post": {
        "tags": [
          "login"
        ],
        "summary": "login",
        "parameters": [
          {
            "in": "body",
            "name": "add request",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/login"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
     "/api/v1/validate": {
      "post": {
        "tags": [
          "validate"
        ],
        "summary": "otp validate",
        "parameters": [
          {
            "in": "body",
            "name": "add request",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/validate"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
     "/api/v1/password_validate": {
      "post": {
        "tags": [
          "validate"
        ],
        "summary": "password validate",
        "parameters": [
          {
            "in": "body",
            "name": "add request",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/password_validate"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
     "/api/v1/forgot_password": {
      "post": {
        "tags": [
          "forgot_password"
        ],
        "summary": "forgot_password",
        "parameters": [
          {
            "in": "body",
            "name": "add request",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/forgot_password"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/search": {
      "get": {
        "tags": [
          "search"
        ],
        "summary": "search",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "type": "string",
            "description": "",
            "required": true
          },
          {
            "in": "query",
            "name": "resource_type",
            "type": "string",
            "description": "",
            "required": true
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/api/v1/user": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "add user",
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "type": "string",
            "description": "",
            "required": true
          },
          {
            "in": "body",
            "name": "add request",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },"get": {
        "tags": [
          "user"
        ],
        "summary": "get user",
        "parameters": [
           {
            "in": "header",
            "name": "Authorization",
            "type": "string",
            "description": "",
            "required": true
          },
          {
            "in": "query",
            "name": "user_email",
            "type": "string",
            "description": "",
            "required": true
          }

        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Ok"
          },
          "400": {
            "description": "Bad request"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }

  },
  "definitions": {
    "login": {
      "type": "object",
      "properties": {
        "user_email": {
          "type": "string",
          "format": "string"
        },
        "password": {
          "type": "string",
          "format": "string"
        }
      }
    },
    "validate": {
      "type": "object",
      "properties": {
        "user_email": {
          "type": "string",
          "format": "string"
        },
        "otp": {
          "type": "string",
          "format": "string"
        }
      }
    },
    "password_validate": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string",
          "format": "string"
        },
        "otp": {
          "type": "string",
          "format": "string"
        }
      }
    },
    "forgot_password": {
      "type": "object",
      "properties": {
        "user_email": {
          "type": "string",
          "format": "string"
        }
      }
    },"user": {
      "type": "object",
      "properties": {
        "user_email": {
          "type": "string",
          "format": "string"
        },
        "password": {
          "type": "string",
          "format": "string"
        }
      }
    }
  },
  "swagger": "2.0"
}